refactor(traces): split Gemini content helpers out of the vertex extractor
Review feedback: vertexAdk.ts exceeded the 300-source-line cap for new
files — the Gemini content conversion helpers move to _geminiContent.ts
(227 source lines remain). Booleans gain is/has prefixes
(hasExtractedParams, hasExtractedUsage, hasSetArgs, hasSetResult), and
the raw-string test context folds into createExtractorContext via a
skipJsonParsing option instead of a parallel helper.
